refactor: refactor landing page layout
