--- src/app/(client)/layout.tsx
@@ -1,9 +1,0 @@
-import Layout from '@/components/ClientLayout';
-
-export default function ClientLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <Layout>{children}</Layout>;
-}
