+++ src/app/layout.tsx
@@ -1,5 +1,8 @@
 import '../styles/global.css';
 
+import { Footer } from '@/components/Layout/Footer';
+import { Header } from '@/components/Layout/Header';
+
 import { notoSans } from './font';
 import Provider from './provider';
 
@@ -18,7 +21,13 @@ export default function RootLayout({
   return (
     <html lang="zh-TW" className={notoSans.className}>
       <body id="app">
-        <Provider>{children}</Provider>
+        <Provider>
+          <div className="flex min-h-screen flex-col overflow-x-hidden">
+            <Header />
+            <main className="grow">{children}</main>
+            <Footer />
+          </div>
+        </Provider>
       </body>
     </html>
   );
